+++ include/fluxions_sphl_sampler.hpp
@@ -31,21 +31,21 @@ namespace Fluxions
 	public:
 		SphlSampler();
 
-		void resize(size_t samplesX, size_t samplesY);
+		void resize(unsigned samplesX, unsigned samplesY);
 		void saveSampleMap(const std::string& path, int pxscale);
 		void sampleCubeMap(const Image4f& cubeMap, MultispectralSph4f& msph);
 
 	private:
-		const size_t pxscale = 4;
+		const unsigned pxscale = 4;
 		Image3i sampleMap;
-		size_t ix = 128;
-		size_t iy = 64;
-		size_t numSamples = 0;
+		unsigned ix = 128;
+		unsigned iy = 64;
+		unsigned numSamples = 0;
 		std::vector<Vector3f> randomVectors;
 		std::vector<double> theta;
 		std::vector<double> phi;
 		std::vector<double> sph;
-		std::vector<size_t> debug_sides;
+		std::vector<unsigned> debug_sides;
 	};
 } // namespace Fluxions
